Always launch docker containers in the correct parent cgroup
Currently, the docker executor script hardcodes the parent cgroup of a docker
container to "mesos/$CGROUPS_GUID". This change makes it instead discover the
cgroup using "/proc/self/cgroup" per the reccommendation of:
  https://www.freedesktop.org/wiki/Software/systemd/PaxControlGroups/

This change also removes the code for setting memory.use_hierarchy=1 because
it is ineffective. In order to set that property, it must be set on the root
cgroup, before any child cgroups are created. My internal changes to cgroups
on our mesos slaves takes care of that and all child cgroups inherit the
property.