Add ability to upload files immediately to S3
Change has two parts: Adds an optional uploadImmediately field to the uploader
metadata, that when present and set will cause the uploader to upload it
immediately on change, and code within the on-detection function
`handleNewOrModifiedS3Metadata` to perform the upload if that field is
detected.

This change in particular involved refactoring the vast majority of the code in
`checkUploads` into a seperate function.  I wanted to avoid duplicating the
logic in `checkUploads` in `handleNewOrModifiedS3Metadata`. If I directly
called `checkUploads`, however, that would upload all the files that were
queued to upload, rather than just the files marked to upload immediately. So I
made a new function that would just take a map of files to upload and upload
only those files. So when it needed to upload the all queued files, as in
`checkUploads`, it could directly work on the `metadataToUploader` field, and
when it only needs to upload a single block of files, it can only be passed
that single uploader/metadata.