Move uploading checks off of fs polling thread
The way that I had written the code for this, the main thread, which was
used to poll the filesystem for metadata changes, was being blocked by
awaiting the files to finish uploading. This moves the code that depends
on the the uploads finishing (namely, logging that it finished and that
closing the uploaders) into callbacks off of the futures.  This change
also no longer places any uploaders into the class-scoped expiredUpload
queue.