Correctly handle existing metadata -> immediate
In general, the upload driver wants to avoid re-adding modified existing
metadata, because it doesn't want to create a new uploader for it and it
doesn't actually (for the most part) look at the contents of the
metadata file, delegating that to the uploader itself. So on the whole
it ignores existing metadata files on change. In this case, in
particular, it needs to pick up changes in the metadata file because it
should still perform the upload asap.

This change just forces the upload if there is an existing metadata file
that has been changed so that it requires immediate upload, it will
upload the indicated files. As a heads-up, if you do this (modify an
existing metadata file), it will still delete the metadata file after it
performs the upload, so if you want to keep that around you will need to
re-create the metadata file.