Replace existing global search
Change from directly using the fuzzy search API provided by `fuzzy` to
the existing custom one used in the search bar on the Requests page.
This did preserve using the existing `react-typeahead` element, rather
than switching to the `react-select` element.

One thing that we lost in this switchover: highlighting of matched
characters. The search on the requests page doesn't use or emit that
data, so it no longer comes into the global search. We also aren't as
yet using the full potential of the global search - for example, you
can't search by the deploy user, and you can't toggle searching by
status.