Add seperate map to track immediate uploaders
Duplicates the same kind of structure that we are currently using to
keep track of the normal loaders with the immediate uploaders.  That is,
I added another map from metadata to immediate uploaders.  This is
checked before creating a new uploader and only cleared when the
existing uploader is definitely done and has been removed.  Hopefully
this should prevent the creation of duplicate immediate uploaders in a
more robust way than the prior attempt, which might not maintain the
block for the lifetime of the first immediate uploader.