Lock test for creating immediate uploader
The way that we are blocking attempts to create duplicate immediate
uploaders is by creating a list of metadata objects that we have
immediate uploaders for, and refusing to create a new uploader if that
list already contains the metadata file associated with the new
uploader. We need to perform the check-and-add operation atomically, to
prevent the creation of two duplicate immediate uploaders to be
interspersed in such a way that they are both created. In order to do
so, this moves the check-and-add operation into a locked block of code
in order to prevent that kind of interspersal.