Mark slaves on inactive host as decommissioned
When a previously-seen slave on a host which is marked as inactive
attempts to join the cluster, it is now marked as `DECOMMISSIONED`.
Previously, it was ignored and nothing actually happened to it. This
actually will stop it from accepting offers, as well as provide
visibility into what is actually going on w/r/t the flapping slave.