Add resource for marking slaves as inactive
Next step toward being able to mark a machine for a slave up-for-review
via the UI. Previously you would have needed to manually edit the node
in ZK in order mark a node as inactive. Notably, un-marking this host as
inactive will not allow the slave to being accepting offers until the
slave is also un-marked as decom'ed or it disappears and reappears with
a new slaveID. So restoring a slave will likely be a two step process.