Fix issues from code review
In particular, it changes a handful of phrasing issues on the UI and
reorganizes how the data is stored in zk. In particular, rather than
single node which contains an array as children, it now has a main node
whose (empty) children represent the hosts that have been deactivated.
There is now an additional method which simply checks whether a host is
active. This allows the query to zk to be deferred until it actually
receives an offer that reveals a new slave.