Add additional hook into component lifecycle
Adds another hook for `rootComponent` to interact with the component
lifecycle. In particular, the `onLoad` function provides a hook that
will run during `componentDidMount`, in contrast to `refresh`, which
provides a hook into `componentWillMount`. The intended use case is for
kicking off actions which should be handled eventually, but which should
not prevent the page from loading. The issue with doing those in
`refresh` is that `rootComponent` will display a spinner until the
promise returned from `refresh` is entirely complete; there isn't a
mechanism for only having some requests block.