Add modal to edit existing priority freeze
Add a button and corresponding modal dialog to adjust the properties of
a priority freeze. Actually, what it does is submits a new priority
freeze that overwrites the existing one, but that's life.

This also fixes an bug in the implementation of the remove priority
freeze action; namely, the 'get priority freeze endpoint' returns an
error (well, a 404) when there is no existing priority freeze. In
general, this doesn't matter, but because the reducer treats it as an
error rather than empty response that part of the state isn't refreshed.
So removing a priority freeze wouldn't update the page to remove the
active-priority-freeze box without refreshing the page. This adds an
action+reducer to clear that part of the page before submitting the
request to deal with that issues.