Transition onLoad to an action dispatcher
One of the changes that was blocked on the tailer was making `onLoad`
behave like an action dispatcher, rather than a function. So each of the
components treats it as an object (or, at least, the base case is `null`
rather than `_.noop`).  This means that it was misbehaving when it
wasn't explicitly set because pre-tailer it was expected to be a
function.