Add spacing line at the bottom of logs
Add a small amount of padding at the bottom of the logging container.
Now if you scroll all the way to the bottom of log tailer, there should
be a small space between the last log line and the bottom of the screen,
so that the scrollbar (on MacOS) won't overlap the last line of the
text.