Don't check length of nullable field
Was previously checking the `length` property to determine if a given
array had contents. This didn't work because in cases where the
directory had been cleaned up, the query to get that field would respond
with an error code, which resulted in `files` being `null` or
`undefined`, rather than the empty array.