Return null when files not unavailable
React requires that a component evaluate to either null, a dom element,
or a react component when it renders. The `TaskLatestLog` component was
rendering as `false`, rather than `null`, in cases when files were not
loaded - in particular, when following links directly into a directory
rather than the task page itself.