Correctly propagate remove request
When a request has active tasks, first the tasks are removed and the
request cleanup request is destroyed. Once all the tasks are removed,
then a new request cleanup request is submitted.  I had missed this
case, so the `removeFromLoadBalancer` flag was set on the first request,
but not on the second, when it would actually take affect. This
correctly propagates the `removeFromLoadBalancer` flag across both
requests.