Reproduce issue causing bounce to get stuck
If a bounce occurs in the quiet period between a task failing and a new
task being launched by Singularity, it seems that the new tasks are
launched and marked as having been started by the bounce, but because no
tasks were ever killed, the bounce isn't marked as finished. Add test
case to reproduce the issue; will work on fix next.