Kill bounce on scheduling new tasks
Shim a fix by removing the bounce lock after scheduling the new tasks
when there is a bounce lock and are currently no active tasks.  This
means that the bounce lock is removed before those tasks become active.
I don't think that this will create any problems; this should be the
only case where there is a bounce present, more instances are being
scheduled, and there are currently no running instances, but it's an odd
enough case that I'm not really sure.