Defer duplicate request processing to scheduler
Defer checking duplicate requests from ZK into the scheduler.  This
allows slightly more control over what kinds of requests are allowed to
be in the pending queue simultaneously, and avoids the kind of edge
cases that we were running into previously by attempting to use
nodenames in ZK to resolve the same problem.

/cc @ssalinas