Respond to PR comments
In particular, make retry more configurable by allowing the user to set
the number of attempts and the strategy to determine whether a response
should be retried. Also explicitly disabled retries on the request, so
that it would be entirely handled within the Guava retryer. The reason
that retries within the HttpClient can't be re-used is that doing so
would just retry the given HttpRequest, and this needs to be able to
alter the HttpRequest between attempts. Also is a little more careful
about retrying to different hosts on requests.