Add test for multipe scheduled requests in queue
Add test to validate how multiple scheduled requests for the same
request are processed in the queue. In particular, validate that when an
immediate request and a scheduled request make it into the pending queue
at the same time, only the immediate request runs and the pending
request gets dropped.