Ignore healthcheck config when skipping checks
Fix deployment failures when deploying to a request with no ports
assigned (and therefore no valid value for healthcheck port index) when
a healthcheck URI is supplied.

Example:

```json
{
  "requestId": "my-test-service",
  "healthcheck": {
    "uri": "/health"
  },
  "resources": {
    "cpus": 0.1,
    "memoryMb": 128,
    "numPorts": 0
  }
}
```

This configuration currently fails even when skipHealthchecks is true.
Ideally Singularity should either

1. Error immediately because supplying healthcheck config for a request
with checks disabled is logically invalid, or

2. Ignore healthcheck metadata entirely when the request isn't going to
use it anyway.

This adds the latter solution, but I believe both are sensible options.